Shrink `subobj_flags` to `uint16_t`
All writes to `subobj_flags` are one of:
- an explicit 0
- a copy of a `subobj_flags` from elsewhere
- bit-or with `(1 << subobj_num)`, where `subobj_num` is a
  `submodel_index`.  Per the comment in the definition of
  `submodel_index`, the valid values for `submodel_index` are `[0, 9]`.
  Therefore, the highest valid value from the shift is `1 << 9`, which
  can be represented in a `uint16_t`.

Switch the type to `uint16_t`, so that `subobj_flags` can be packed
immediately after `anim_angles`.

+++ common/main/object.h
@@ -369,7 +369,7 @@ struct polyobj_info : prohibit_void_ptr<polyobj_info>
 {
 	polygon_model_index model_num{};// which polygon model
 	std::array<vms_angvec, MAX_SUBMODELS> anim_angles{}; // angles for each subobject
-	int subobj_flags{0};       // specify which subobjs to draw
+	uint16_t subobj_flags{0};       // specify which subobjs to draw
 	int tmap_override{0};      // if this is not -1, map all face to this
 	int alt_textures{0};       // if not -1, use these textures instead
 };
